+++ fealpy/vem/__init__.py
@@ -1,22 +1,5 @@
-from .scaled_monomial_space_mass_integrator import ScaledMonomialSpaceMassIntegrator2d
-from .scaled_monomial_space_diffusion_integrator import ScaledMonomialSpaceDiffusionIntegrator2d 
-
-from .conforming_vem_dof_integrator import ConformingVEMDoFIntegrator2d
-from .conforming_scalar_vem_h1_projector import ConformingScalarVEMH1Projector2d
-from .conforming_scalar_vem_l2_projector import ConformingScalarVEML2Projector2d
-from .conforming_scalar_vem_mass_projector import ConformingScalarVEMMassIntegrator2d
-
-#from .non_conforming_vem_dof_integrator import NonConformingVEMDoFIntegrator2d
-#from .non_conforming_scalar_vem_h1_projector import NonConformingScalarVEMH1Projector2d
-#from .non_conforming_scalar_vem_l2_projector import NonConformingScalarVEML2Projector2d
-
-from .laplace_integrator import ConformingScalarVEMLaplaceIntegrator2d
-#from .laplace_integrator import  NonConformingScalarVEMLaplaceIntegrator2d
-from .scalar_source_integrator import ConformingVEMScalarSourceIntegrator2d 
-#from .scalar_source_integrator import NonConformingVEMScalarSourceIntegrator2d
-from .dof_dof_stabilization_term_integrator import DofDofStabilizationTermIntegrator2d
 from .bilinear_form import BilinearForm
 from .linear_form import LinearForm
-from .estimator import PoissonCVEMEstimator
 
-#from .poisson_vem_adaptive_solver import PoissonACVEMSolver
+from .scalar_diffusion_integrator import ScalarDiffusionIntegrator
+from .scalar_source_integrator import ScalarSourceIntegrator
