+++ fealpy/vem/scalar_source_integrator.py
@@ -1,74 +1,66 @@
+from typing import Optional, Literal                                            
+                                                                                
+from ..backend import backend_manager as bm                                     
+from ..typing import TensorLike, Index, _S, SourceLike                          
+                                                                                
+from ..mesh import HomogeneousMesh                                              
+from ..functionspace.space import FunctionSpace as _FS                          
+from ..utils import process_coef_func                                           
+from ..functional import linear_integral                                        
+from .integrator import LinearInt, SrcInt, CellInt, enable_cache, assemblymethod
+                                                                                
+                                                                                
+class ScalarSourceIntegrator(LinearInt, SrcInt, CellInt):                       
+    r"""The domain source integrator for function spaces based on homogeneous meshes."""
+    def __init__(self, source: Optional[SourceLike]=None, q: int=None, *,          
+                 region: Optional[TensorLike] = None,                           
+                 batched: bool=False,                                           
+                 method: Literal['isopara', None] = None) -> None:              
+        super().__init__(method=method if method else 'assembly')               
+        self.source = source                                                    
+        self.q = q                                                              
+        self.set_region(region)                                                 
+        self.batched = batched                                                  
+                                                                                
+    @enable_cache                                                               
+    def to_global_dof(self, space: _FS, /, indices=None) -> TensorLike:         
+        if indices is None:                                                     
+            return space.cell_to_dof()                                          
+        return space.cell_to_dof(index=self.entity_selection(indices))          
+                                                                                
+    @enable_cache                                                               
+    def fetch(self, space: _FS, /, inidces=None):                               
+        q = self.q                                                              
+        index = self.entity_selection(inidces)                                  
+        mesh = getattr(space, 'mesh', None)                                     
+                                                                                
+        if not isinstance(mesh, HomogeneousMesh):                               
+            raise RuntimeError("The ScalarSourceIntegrator only support spaces on"
+                               f"homogeneous meshes, but {type(mesh).__name__} is"
+                               "not a subclass of HomoMesh.")                   
+                                                                                
+        cm = mesh.entity_measure('cell', index=index)                           
+        q = space.p+3 if self.q is None else self.q                             
+        qf = mesh.quadrature_formula(q, 'cell')                                 
+        bcs, ws = qf.get_quadrature_points_and_weights()                        
+        phi = space.basis(bcs, index=index)                                     
+        return bcs, ws, phi, cm, index                                   
 
-import inspect
+    def assembly(self, space):
+        f = self.source
 
-import numpy as np
-from numpy.typing import NDArray
-from typing import TypedDict, Callable, Tuple, Union
+        phi = space.smspace.basis                                               
+        p = space.p                                                             
+        q = p+3                                                                 
+        def u(x, index):                                                        
+            return bm.einsum('ij, ijm->ijm', f(x), phi(x, index))               
+        bb = space.mesh.integral(u, q=q, celltype=True)                         
+        g = lambda x: x[0].T @ x[1]                                             
+        PI0 = space.PI0                        
+        bb = bm.concatenate(list(map(g, zip(PI0, bb))))                         
+        return bb 
 
-class ConformingVEMScalarSourceIntegrator2d():
-    def __init__(self, f: Union[Callable, int, float, NDArray], PI0, c=None):
-        """
-        @brief
 
-        @param[in] f 
-        """
-        self.coef = 1 if c is None else c
-        self.f = f
-        self.vector = None
-        self.PI0 = PI0
 
-    def assembly_cell_vector(self, space, index=np.s_[:], cellmeasure=None, out=None, q=None):
-        """
-        @brief 组装单元向量
 
-        @param[in] space 一个标量的函数空间
 
-        """
-        f = self.f
-        N = len(inspect.signature(f).parameters)
-        phi = space.smspace.basis
-        p = space.p
-        q = p + 3 if q is None else q
-        if N==1:
-            def u(x, index):
-                return np.einsum('ij, ijm->ijm', f(x), phi(x, index=index))
-        else:
-            def u(x, index):
-                return np.einsum('ij, ijm->ijm', f(x, index), phi(x, index=index))
-        bb = space.mesh.integral(u, q=q, celltype=True)
-        if isinstance(self.coef, (int, float)):
-            g = lambda x: x[0].T@x[1]
-            bb = np.concatenate(list(map(g, zip(self.PI0, bb))))
-            return self.coef*bb
-        elif isinstance(self.coef, np.ndarray):
-            g = lambda x: (x[0].T@x[1])*x[2]
-            bb = np.concatenate(list(map(g, zip(self.PI0, bb, self.coef))))
-        return bb
-
-class NonConformingVEMScalarSourceIntegrator2d():
-    def __init__(self, f: Union[Callable, int, float, NDArray], PI0):
-        """
-        @brief
-
-        @param[in] f 
-        """
-        self.f = f
-        self.vector = None
-        self.PI0 = PI0
-
-    def assembly_cell_vector(self, space, index=np.s_[:], cellmeasure=None, out=None, q=None):
-        """
-        @brief 组装单元向量
-
-        @param[in] space 一个标量的函数空间
-
-        """
-        f = self.f
-        phi = space.smspace.basis
-        p = space.p
-        def u(x, index):
-            return np.einsum('ij, ijm->ijm', f(x), phi(x, index=index))
-        bb = space.mesh.integral(u, q=p+3, celltype=True)
-        g = lambda x: x[0].T@x[1]
-        bb = np.concatenate(list(map(g, zip(self.PI0, bb))))
-        return bb
