--- fealpy/vem/vem/__int__.py
@@ -1,5 +1,0 @@
-from .bilinear_form import BilinearForm
-from .linear_form import LinearForm
-
-from .scalar_diffusion_integrator import ScalarDiffusionIntegrator
-from .scalar_source_integrator import ScalarSourceIntegrator
